chore(authn): all jsons and env files required for init are at the same directory level as docker-compose
