+++ authn/server_details.json
@@ -1,0 +1,9 @@
+{
+  "service": "authentication",
+  "build": "multi_purpose_single_server",
+  "version": {
+    "major": 0,
+    "minor": 3,
+    "build": 0
+  }
+}
