+++ authn/webapi/server/details/details_test.go
@@ -1,0 +1,14 @@
+package details
+
+import (
+	"testing"
+)
+
+
+func TestReadDetailsFromFile(t *testing.T) {
+	details := ReadDetailsFromFile()
+
+	if details == nil {
+		t.Error("server details are nil")
+	}
+}
